rename options to body
