+++ lib/src/models/command_palette_config.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:command_palette/src/controller/command_palette_controller.dart';
-import 'package:command_palette/src/widgets/options/command_palette_options.dart';
+import 'package:command_palette/src/widgets/options/command_palette_body.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
