+++ lib/src/widgets/options/command_palette_body.dart
